backend/src/main/java/edu/ap/projectteambisfits/role/RoleServiceImpl.java
package edu.ap.projectteambisfits.role;

import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class RoleServiceImpl implements RoleService {

    @Autowired
    private RoleRepository roleRepository;

    @Override
    public List<Role> findAll() {
        return roleRepository.findAll();
    }

    @Override
    public void deleteRole(Role role) {
        roleRepository.delete(role);
    }

    @Override
    public Role saveRole(Role role) {
        return roleRepository.save(role);
    }
}